requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# HTTP client for external API calls
aiohttp==3.9.1
requests==2.31.0

# AI and content generation
openai==1.102.0
httpx==0.27.2

# Browser automation
playwright==1.40.0

# Async support
asyncio-mqtt==0.16.1

# Data processing
python-dateutil==2.8.2

# Environment and configuration
python-dotenv==1.0.0

# Cryptography for secure data handling
cryptography==41.0.7

# JSON Web Tokens for authentication
PyJWT==2.8.0

# PostgreSQL support
psycopg2-binary==2.9.9

# Socket.IO async support for production
eventlet==0.33.3
